pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reasoning-router"
version = "0.1.0"
description = "Dynamic Multi-Strategy Reasoning for Robust Multi-Step Problem Solving"
authors = [
    {name = "Praneeth Vadlapati", email = "praneeth.vad@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "datasets",
    "langgraph",
    "langgraph-prebuilt",
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "langchain[evaluation]",
	"numpy",
	"optuna",
    "python-dotenv",
	"torch",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["reasoning_router*"]
